Fix more printf type warnings

+++ firmware/src/gui/helpers/lvgl_mem_helper.hh
@@ -9,7 +9,7 @@ struct LVGLMemory {
 		lv_mem_monitor(&mon);
 		printf_("%s: used: %6d (%3d %%), frag: %3d %%, biggest free: %6d\n\r",
 				tag.data(),
-				(int)mon.total_size - mon.free_size,
+				(int)mon.total_size - (int)mon.free_size,
 				mon.used_pct,
 				mon.frag_pct,
 				(int)mon.free_biggest_size);
